[김진원] 좌석선택 페이지 기능적 간이구현(1)

+++ D_airline/src/main/java/dao/SeatDAO.java
@@ -166,7 +166,7 @@ public class SeatDAO {
 			m.put("departureTime", rs.getString("departureTime"));
 			m.put("arrivalTime", rs.getString("arrivalTime"));
 			m.put("flightStatus", rs.getString("flightStatus"));
-			m.put("seatPrice", rs.getInt("seatPrice"));
+			m.put("seatPrice", rs.getDouble("seatPrice"));
